requirements.txt
# Lightweight dependencies for basic video search
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
imagehash>=4.3.1
tqdm>=4.65.0
matplotlib>=3.7.0
scipy>=1.11.0

# Lightweight text encoder
sentence-transformers>=2.2.2

# Note: This lightweight version doesn't require:
# - transformers (for BLIP)
# - CLIP-by-OpenAI (uses sentence transformers instead)
# - segment-anything (no advanced object detection)
# - grad-cam (no visual explanations)
